third_party/WebKit/LayoutTests/http/tests/inspector-enabled/console-clear-arguments-on-frame-navigation-expected.txt

CONSOLE MESSAGE: line 2: A message with first argument string
CONSOLE MESSAGE: line 3: 2011
CONSOLE MESSAGE: line 4: [object Window]
Tests that Web Inspector will discard console message arguments and convert first argument into a string when iframe where the message was logged is navigated to a different page.


Message: "A message with first argument string", arguments: [string]
Message: "2011", arguments: [string]
Message: "[object Window]", arguments: [string]
